# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks/clerk:
    post:
      operationId: clerkWebhook
      tags: [webhooks]
      summary: Clerk webhook handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces for the current user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace with members
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceWithMembers"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [workspaces]
      summary: List members of a workspace
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: List of workspace members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks in a project
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (any workspace member)
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task with comments
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (assignee, team_lead, or ADMIN)
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (assignee, team_lead, or ADMIN)
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        image:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, image, createdAt, updatedAt]
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        ownerId:
          type: string
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, ownerId, imageUrl, createdAt, updatedAt]
    WorkspaceMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        workspaceId:
          type: string
        role:
          type: string
          enum: [ADMIN, MEMBER]
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, workspaceId, role, message, user, createdAt]
    WorkspaceWithMembers:
      allOf:
        - $ref: "#/components/schemas/Workspace"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/WorkspaceMember"
          required: [members]
...
    TaskDetail:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        type:
          type: string
          enum: [TASK, BUG, FEATURE, IMPROVEMENT, OTHER]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        assigneeId:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        assignee:
          $ref: "#/components/schemas/User"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentDetail"
      required: [id, projectId, title, status, type, priority, assigneeId, dueDate, createdAt, updatedAt, assignee, comments]
...
    CommentDetail:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        userId:
          type: string
        taskId:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required: [id, content, userId, taskId, createdAt, user]
...
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        type:
          type: string
          enum: [TASK, BUG, FEATURE, IMPROVEMENT, OTHER]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        assigneeId:
          type: string
        dueDate:
          type: string
          format: date-time
      required: [title, assigneeId, dueDate]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        type:
          type: string
          enum: [TASK, BUG, FEATURE, IMPROVEMENT, OTHER]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        assigneeId:
          type: string
        dueDate:
          type: string
          format: date-time
...