# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";
import type { UpdateTaskBodyType } from "./updateTaskBodyType";

export interface UpdateTaskBody {
  title?: string;
  description?: string;
  status?: UpdateTaskBodyStatus;
  type?: UpdateTaskBodyType;
  priority?: UpdateTaskBodyPriority;
  assigneeId?: string;
  dueDate?: Date;
}